Containerize Header
In order to make the Login responsive to the state it was necessary
or at least expedient to create a container component for the
Header.
