+++ src/client/actions/settings.js
@@ -4,6 +4,9 @@ export const GET_SETTINGS = 'GET_SETTINGS'
 export const SET_SETTINGS = 'SET_SETTINGS'
 export const UPDATE_SETTINGS = 'UPDATE_SETTINGS'
 export const SAVE_SETTINGS = 'SAVE_SETTINGS'
+export const GET_USER = 'GET_USER'
+export const SET_USER = 'SET_USER'
+
 
 export const setSettings = (appKey, appSecret) => {
   return {
@@ -42,5 +45,22 @@ export const saveSettings = () => {
     fetch('/api/v1/settings', opts)
       .then((resp) => resp.json())
       .then(() => dispatch(push('/')))
+  }
+}
+
+export const setUser = (user) => {
+  return {
+    type: SET_USER,
+    twitterUsername: user.username
+  }
+}
+
+export const getUser = () => {
+  return (dispatch) => {
+    fetch('/api/v1/user', {credentials: 'same-origin'})
+      .then((resp) => resp.json())
+      .then((result) => {
+        dispatch(setUser(result.user))
+      })
   }
 }
