+++ src/client/components/Login.js
@@ -1,10 +1,16 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import { Link } from 'react-router-dom'
+import { getUser } from '../actions/settings'
+import store from '../store'
 
 export default class Login extends Component {
 
   // <img title={ this.props.username } id="avatar" src={ this.props.avatar_url } /> &nbsp;
+
+  componentWillMount() {
+    store.dispatch(getUser())
+  }
 
   render() {
     if (this.props.username) {
