+++ src/client/reducers/settings.js
@@ -1,4 +1,4 @@
-import { SET_SETTINGS, UPDATE_SETTINGS } from '../actions/settings'
+import { SET_SETTINGS, UPDATE_SETTINGS, SET_USER } from '../actions/settings'
 
 const initialState = {
   appKey: '',
@@ -22,6 +22,13 @@ export default function settings(state = initialState, action) {
       }
       s[action.name] = action.value
       return s
+    }
+
+    case SET_USER: {
+      return {
+        ...state,
+        twitterUsername: action.twitterUsername
+      }
     }
 
     default: {
