--- t.js
@@ -1,3 +1,0 @@
-import { addKeys } from './src/server/auth'
-
-console.log(addKeys)
